+++ src/main/java/tracks/TrackIntervalFeature.java
@@ -152,9 +152,6 @@ public class TrackIntervalFeature extends Track {
 					break;
 				}
 				IntervalFeature intervalFeature= new IntervalFeature(q, this.getTrackFormat(), null);
-//				if(intervalFeature.getRaw().contains("\t__ignore_me__")){ // Hack to circumvent issue #38
-//					continue;
-//				}
 				xFeatures.add(intervalFeature);
 			} 
 		}
@@ -192,9 +189,6 @@ public class TrackIntervalFeature extends Track {
 				break;
 			}
 			IntervalFeature intervalFeature= new IntervalFeature(q, TrackFormat.VCF, this.getVCFCodec());
-//			if(q.contains("\t__ignore_me__")){ // Hack to circumvent issue #38
-//				continue;
-//			}
 			xFeatures.add(intervalFeature);
 		} 
 		return xFeatures;
@@ -207,13 +201,9 @@ public class TrackIntervalFeature extends Track {
 	 * */
 	protected Boolean featureIsVisible(String x) throws InvalidGenomicCoordsException, IOException{
 		
-//		if(x.contains("__ignore_me__")){
-//			return false;
-//		}
-		
 		boolean showIt= true;
 		if(this.getShowRegex() != null && 
-		   ! this.getShowRegex().equals(Filter.DEFAULT_SHOW_REGEX.getValue())){
+		   ! this.getShowRegex().equals(Pattern.compile(Filter.DEFAULT_SHOW_REGEX.getValue()))){
 			showIt= this.getShowRegex().matcher(x).find();
 		}